Fix a bug and make the lichen boundry stable.
If a tile is already expanded, it should be added to the positions to water unconditionally. However, the current water info computation did not, and results in a strange behavior of lichen growth -- the unstable boundary. For example. consider following case on a 1x4 map. There are two lichen strains, A and B. A and B initially occupies tile 1 and 4. In first turn, they both water, then they expanded to cell 2 and cell 3 respectively. Next turn, they both water again. However, they lost cell 2 and cell 3 respectively. This is because cell 2 is next to cell 3, which belongs to another strain, so it cannot be watered by A. Similarly, cell 3 cannot be watered by B. Finally, A lost cell 2 and B lost cell 3. If A and B keep watering, the two pattern will periodically appear, resulting in a unstable boundary, even if A and B belongs to the same player. 

```
step 1: A _ _ A
step 2: A A B B
step 3: A _ _ A
...
```

In short, if a tile is already expanded and reachable from the factory, it should be added to the positions to water unconditionally, no matter whether its neighbors belong to other strains or not.