Fix a bug in env.step that may lead to crashing.
Fix a bug in env.step that may lead to crashing. When updating the action queue, all exceptions will be caught and printed. However,  `e.with_traceback()` takes one argument, and none is given. This will lead to the whole game crashing. When a player provides actions for a non-existing robot or a robot he/she does not own, the crash is triggered.