Fix a bug in compute_water_info
Fix a bug in compute_water_info

The bug leads to unexpected connection between lichen strains.
For example, on a 4x1 map. There is a lichen strain A. The third tile of lichen has been digged out by enemy robots. However, with the bug, the two part of lichen strains are still considered as connected, and will be watered together.

```
|A|
|A|
| |
|A|
```