Fix: make transfer action irrelevant to unit_id.
Before, the transfer result relies on the unit_id because the transfer is executed from units with small id to ones with big id. For example, consider three light robots, A, B, and C, with a cargo space of 100. Initially, they both have 80 water. Now, A wants to transfer 80 water to B, and B wants to transfer 80 water to C.
```
           A   B   C
initial:  80  80  80

action: A -80-> B -80-> C
```
If id(A) < id(B), then A transfers first. After transfer, B has 100 water because of the cargo space limit. Then B transfers 80 water to C and has only 20 left.

If id(A) > id(B), then B transfer first. After transfer, B has 0 water. Then A transfer 80 water to B, and B finally has 100.
```
          A   B   C
final:    0  20 100, if id(A) < id(B)
          0  80 100, if id(A) > id(B)
```

Conceptually, all transfer actions happen simultaneously and should not rely on the unit id. This commit fixes this issue and implements the id-irrelevant transfer. In this commit, we first take away resources/power that is going to be transferred from all robots before adding resources to any target. Then we add them to targets altogether and perform the cargo space limit checks.