fix: actions correctly repeat now
The change log of v2.0.0 says, "actions in an action queue can specify both the number of times to repeat it directly, as well as whether to put it back to the end of the action queue once exhausted." However, it is not implemented in `unit.repeat_action`. This commit implements the such feature.